+++ src/c2goto/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_executable(c2goto c2goto.cpp)
 target_include_directories(c2goto
-    PRIVATE ${Boost_INCLUDE_DIRS} 
+    PRIVATE ${Boost_INCLUDE_DIRS}
 )
-target_link_libraries(c2goto ansicfrontend clangcfrontend cppfrontend gotoprograms langapi  cprover_blank_library "${Boost_LIBRARIES}")
+target_link_libraries(c2goto ansicfrontend clangcfrontend cppfrontend clangcppfrontend gotoprograms langapi  cprover_blank_library "${Boost_LIBRARIES}")
 
 if (DEFINED HAVE_PYTHON)
   target_link_libraries(c2goto -lm -ldl ${BOOST_PYTHON} ${PYTHON_LIBRARIES}   dl util expat z pthread)
@@ -60,11 +60,11 @@ message(STATUS "Multiarch include for c2goto: ${multiarch}. You may want to over
 mangle_clib(mangled_files ${inputs})
 add_library(clibs ${mangled_files} cprover_library.cpp)
 target_include_directories(clibs
-    PRIVATE ${Boost_INCLUDE_DIRS} 
+    PRIVATE ${Boost_INCLUDE_DIRS}
 )
 add_library(cprover_blank_library cprover_blank_library.cpp)
-target_include_directories(cprover_blank_library 
+target_include_directories(cprover_blank_library
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
-    PRIVATE ${Boost_INCLUDE_DIRS} 
+    PRIVATE ${Boost_INCLUDE_DIRS}
 )
